Move interpolation mode enum and functions to separate file
